+++ CHANGELOG.md
@@ -48,6 +48,9 @@ Added:
 
 - Added support for saving smart contract template params in ASCCache.
 
+- Teal v7 support:
+  - opcode `base64decode` ([##653](https://github.com/scale-it/algo-builder/pull/653))
+
 ### Template improvements
 
 - We updated the examples/DAO design. We removed treasury Smart Signature to simplify deposit management. Now a DAO app is managing voting, deposits and treasury.